+++ internal/core/ports/account_repository.go
@@ -10,7 +10,7 @@ import (
 // This is a port in hexagonal architecture - it defines WHAT we need without HOW
 type AccountRepository interface {
 	Create(ctx context.Context, account *domain.Account) (*domain.Account, error)
-	FindByID(ctx context.Context, id int) (*domain.Account, error)
+	FindByID(ctx context.Context, id int64) (*domain.Account, error)
 	FindByDocumentNumber(ctx context.Context, documentNumber string) (*domain.Account, error)
 	GetAll(ctx context.Context) ([]*domain.Account, error)
 }
