+++ internal/core/ports/operation_type_repository.go
@@ -8,7 +8,7 @@ import (
 
 // OperationTypeRepository defines the interface for operation type data operations
 type OperationTypeRepository interface {
-	FindByID(ctx context.Context, id int) (*domain.OperationType, error)
+	FindByID(ctx context.Context, id int64) (*domain.OperationType, error)
 	GetAll(ctx context.Context) ([]*domain.OperationType, error)
 	// Seed initializes the database with the predefined operation types - This should be called during application startup
 	Seed(ctx context.Context) error
